Reorganize DNF libraries into dnflibs package
Consolidate DNF-related libraries (dnfconfig, dnfplugin, module) into a
new dnflibs package for better organization. The old module locations
remain as deprecated shims using @deprecated decorator, ensuring full
backward compatibility while encouraging migration to the new structure.

I plan additional changes in future and I consider the original
structure problematic, like:
* I want to expose some of private functions, but there is no good
  place where to put them - and keeping them in original places is
  not good.
* The module.py library has kind of ambiguous name and it was not
  clear what to expect from it without reading the content.

With this change, all actions related to DNF are nicely closured.

Note that unit-tests are going to be added in a following commits.

Changes:
- Create repos/system_upgrade/common/libraries/dnflibs/ package
- Move dnfconfig.py to dnflibs/dnfconfig.py
- Move dnfplugin.py to dnflibs/dnfplugin.py
- Move module.py    to dnflibs/dnfmodule.py (renamed)
- Mark original libraries and their content as deprecated
- Update the documentation to reflect deprecated functionality
