+++ repos/system_upgrade/common/libraries/dnflibs/__init__.py
@@ -1,0 +1,8 @@
+"""
+DNF-related libraries for the upgrade process.
+
+This package consolidates DNF functionality previously scattered across:
+- leapp.libraries.common.dnfconfig -> dnflibs.dnfconfig
+- leapp.libraries.common.dnfplugin -> dnflibs.dnfplugin
+- leapp.libraries.common.module -> dnflibs.dnfmodule
+"""
